requirements.txt
asgiref==3.11.0
Django==6.0.1
pillow==12.1.0
sqlparse==0.5.5
google-generativeai==0.8.3
